factor out source management
